장바구니 목록 구현 및 테스트

+++ src/main/java/com/spare/dogiver/domain/Cart.java
@@ -2,9 +2,6 @@ package com.spare.dogiver.domain;
 
 import java.util.Date;
 
-
-
-
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,19 +14,19 @@ public class Cart {
 	private Member member;
 	private Goods goods;
 	
-	private long cartCnt;
+	private int cartCnt;
 	
 	private Date createdDate;
 	private Date modifedDate;
 	
 	@Builder
-	public Cart(Member member, Goods goods, long cartCnt) {
+	public Cart(Member member, Goods goods, int cartCnt) {
 		this.member = member;
 		this.goods = goods;
 		this.cartCnt = cartCnt;
 	}
 	
-	public void update(long cartCnt) {
+	public void update(int cartCnt) {
 		this.cartCnt = cartCnt;
 	}
 }
